Linux port: Fix TE/campaign unit-decode desync (MissionRequestClass ulong flags)
THE root cause of TE missions failing to load (and corrupt-entity crashes in the
ones that did): MissionRequestClass had a `ulong flags` member - 8 bytes on
64-bit Linux vs 4 on 32-bit Windows. PackageClass (de)serializes a
MissionRequestClass as a blob (memcpychk(&mis_request, stream,
sizeof(MissionRequestClass), rem)), so the larger Linux struct (8-byte flags +
the 8-byte alignment padding it forces) read past the 32-bit Windows file
layout, desyncing the unit stream: the package over-read ~12 bytes, the next
unit (a flight) read garbage and threw InvalidBufferException.

Fix: ulong flags -> uint32_t flags (the flag values fit in 32 bits). The struct
now matches the Windows 4-byte layout. Verified: the landing/instrument-landing
TE missions now decode fully - all 3 units consume exactly to rem=0,
LoadCampaign() returns 1, FM_JOIN_SUCCEEDED, no crash. This also fixes the
"decodes but corrupts entity data" missions (bad team indices etc. that drove
the recurring TeamInfo[OOB]/dangling-pointer crashes).

+++ src/campaign/include/mission.h
@@ -1,3 +1,4 @@
+#include <cstdint>  // FF_LINUX: uint32_t
 /**************************************************************************
 *
 * Mission.h
@@ -339,7 +340,10 @@ public:
     Team vs;
     CampaignTime tot;
     GridIndex tx, ty;
-    ulong flags;
+    uint32_t flags;  // FF_LINUX: was 'ulong' (8 bytes on 64-bit) - broke the
+                     // blob (de)serialization of MissionRequestClass vs the
+                     // 32-bit Windows file layout, desyncing TE/campaign unit
+                     // decode (PackageClass reads sizeof(MissionRequestClass)).
     short caps;
     short target_num;
     short speed;
